+++ monkey/monkey_island/cc/resources/ransomware_report.py
@@ -2,10 +2,12 @@ import flask_restful
 from flask import jsonify
 
 from monkey_island.cc.resources.auth.auth import jwt_required
-from monkey_island.cc.services.ransomware_report import get_propagation_stats
+from monkey_island.cc.services.ransomware import ransomware_report
 
 
 class RansomwareReport(flask_restful.Resource):
     @jwt_required
     def get(self):
-        return jsonify({"report": None, "propagation_stats": get_propagation_stats()})
+        return jsonify(
+            {"report": None, "propagation_stats": ransomware_report.get_propagation_stats()}
+        )
