+++ monkey/monkey_island/cc/services/ransomware/ransomware_report.py
